Minor improvements and test renaming
