+++ kora-java-crud-submodule/kora-java-crud-submodule-app/src/test/java/ru/tinkoff/kora/example/submodule/app/BlackBoxTests.java
@@ -1,16 +1,13 @@
-package ru.tinkoff.kora.example.graalvm.crud.cassandra;
+package ru.tinkoff.kora.example.submodule.app;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 import io.goodforgod.testcontainers.extensions.ContainerMode;
 import io.goodforgod.testcontainers.extensions.Network;
-import io.goodforgod.testcontainers.extensions.cassandra.CassandraConnection;
-import io.goodforgod.testcontainers.extensions.cassandra.ConnectionCassandra;
-import io.goodforgod.testcontainers.extensions.cassandra.Migration;
-import io.goodforgod.testcontainers.extensions.cassandra.TestcontainersCassandra;
-import io.goodforgod.testcontainers.extensions.redis.ConnectionRedis;
-import io.goodforgod.testcontainers.extensions.redis.RedisConnection;
-import io.goodforgod.testcontainers.extensions.redis.TestcontainersRedis;
+import io.goodforgod.testcontainers.extensions.jdbc.ConnectionPostgreSQL;
+import io.goodforgod.testcontainers.extensions.jdbc.JdbcConnection;
+import io.goodforgod.testcontainers.extensions.jdbc.Migration;
+import io.goodforgod.testcontainers.extensions.jdbc.TestcontainersPostgreSQL;
 import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
@@ -18,49 +15,36 @@ import java.time.Duration;
 import java.util.Map;
 import org.json.JSONObject;
 import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.skyscreamer.jsonassert.JSONAssert;
 import org.skyscreamer.jsonassert.JSONCompareMode;
 
-@TestcontainersCassandra(
+@TestcontainersPostgreSQL(
         network = @Network(shared = true),
         mode = ContainerMode.PER_RUN,
         migration = @Migration(
-                locations = "migrations",
-                engine = Migration.Engines.SCRIPTS,
+                engine = Migration.Engines.FLYWAY,
                 apply = Migration.Mode.PER_METHOD,
                 drop = Migration.Mode.PER_METHOD))
-@TestcontainersRedis(
-        network = @Network(shared = true),
-        mode = ContainerMode.PER_RUN)
-class GraalVMPetControllerTests {
+class BlackBoxTests {
 
     private static final AppContainer container = AppContainer.build()
             .withNetwork(org.testcontainers.containers.Network.SHARED);
 
-    @ConnectionCassandra
-    private CassandraConnection connection;
+    @ConnectionPostgreSQL
+    private JdbcConnection connection;
 
-    @BeforeEach
-    public void setup(@ConnectionCassandra CassandraConnection cassandraConnection,
-                      @ConnectionRedis RedisConnection redisConnection) {
-        if (!container.isRunning()) {
-            var paramsCassandra = cassandraConnection.paramsInNetwork().orElseThrow();
-            var paramsRedis = redisConnection.paramsInNetwork().orElseThrow();
-            container.withEnv(Map.of(
-                    "CASSANDRA_CONTACT_POINTS", paramsCassandra.contactPoint(),
-                    "CASSANDRA_USER", paramsCassandra.username(),
-                    "CASSANDRA_PASS", paramsCassandra.password(),
-                    "CASSANDRA_DC", paramsCassandra.datacenter(),
-                    "CASSANDRA_KEYSPACE", paramsCassandra.keyspace(),
-                    "CACHE_EXPIRE_WRITE", "0s",
-                    "REDIS_URL", paramsRedis.uri().toString(),
-                    "REDIS_USER", paramsRedis.username(),
-                    "REDIS_PASS", paramsRedis.password()));
+    @BeforeAll
+    public static void setup(@ConnectionPostgreSQL JdbcConnection connection) {
+        var params = connection.paramsInNetwork().orElseThrow();
+        container.withEnv(Map.of(
+                "POSTGRES_JDBC_URL", params.jdbcUrl(),
+                "POSTGRES_USER", params.username(),
+                "POSTGRES_PASS", params.password(),
+                "CACHE_EXPIRE_WRITE", "0s"));
 
-            container.start();
-        }
+        container.start();
     }
 
     @AfterAll
@@ -89,6 +73,7 @@ class GraalVMPetControllerTests {
 
         // then
         connection.assertCountsEquals(1, "pets");
+        connection.assertCountsEquals(1, "categories");
         var responseBody = new JSONObject(response.body());
         assertNotNull(responseBody.query("/id"));
         assertNotEquals(0L, responseBody.query("/id"));
@@ -96,6 +81,23 @@ class GraalVMPetControllerTests {
         assertEquals(requestBody.query("/name"), responseBody.query("/name"));
         assertNotNull(responseBody.query("/category/id"));
         assertEquals(requestBody.query("/category/name"), responseBody.query("/category/name"));
+    }
+
+    @Test
+    void getPetNotFound() throws Exception {
+        // given
+        var httpClient = HttpClient.newHttpClient();
+
+        // when
+        var getRequest = HttpRequest.newBuilder()
+                .GET()
+                .uri(container.getURI().resolve("/v3/pets/1"))
+                .timeout(Duration.ofSeconds(5))
+                .build();
+
+        // then
+        var getResponse = httpClient.send(getRequest, HttpResponse.BodyHandlers.ofString());
+        assertEquals(404, getResponse.statusCode(), getResponse.body());
     }
 
     @Test
@@ -117,6 +119,7 @@ class GraalVMPetControllerTests {
         var createResponse = httpClient.send(createRequest, HttpResponse.BodyHandlers.ofString());
         assertEquals(200, createResponse.statusCode(), createResponse.body());
         connection.assertCountsEquals(1, "pets");
+        connection.assertCountsEquals(1, "categories");
         var createResponseBody = new JSONObject(createResponse.body());
 
         // then
@@ -151,6 +154,7 @@ class GraalVMPetControllerTests {
         var createResponse = httpClient.send(createRequest, HttpResponse.BodyHandlers.ofString());
         assertEquals(200, createResponse.statusCode(), createResponse.body());
         connection.assertCountsEquals(1, "pets");
+        connection.assertCountsEquals(1, "categories");
         var createResponseBody = new JSONObject(createResponse.body());
 
         // when
@@ -202,6 +206,7 @@ class GraalVMPetControllerTests {
         var createResponse = httpClient.send(createRequest, HttpResponse.BodyHandlers.ofString());
         assertEquals(200, createResponse.statusCode(), createResponse.body());
         connection.assertCountsEquals(1, "pets");
+        connection.assertCountsEquals(1, "categories");
         var createResponseBody = new JSONObject(createResponse.body());
 
         // when
