+++ kora-java-crud-submodule/kora-java-crud-submodule-pet-api/src/main/resources/reference.conf
@@ -19,6 +19,7 @@ resilient {
     delay = "500ms"
     delayStep = "5s"
     attempts = 3
+    attempts = ${?RETRY_ATTEMPTS}
   }
 }
