+++ kora-java-graalvm-crud-cassandra/src/test/java/ru/tinkoff/kora/example/graalvm/crud/cassandra/BlackBoxTests.java
@@ -1,13 +1,16 @@
-package ru.tinkoff.kora.example.submodule.app;
+package ru.tinkoff.kora.example.graalvm.crud.cassandra;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 import io.goodforgod.testcontainers.extensions.ContainerMode;
 import io.goodforgod.testcontainers.extensions.Network;
-import io.goodforgod.testcontainers.extensions.jdbc.ConnectionPostgreSQL;
-import io.goodforgod.testcontainers.extensions.jdbc.JdbcConnection;
-import io.goodforgod.testcontainers.extensions.jdbc.Migration;
-import io.goodforgod.testcontainers.extensions.jdbc.TestcontainersPostgreSQL;
+import io.goodforgod.testcontainers.extensions.cassandra.CassandraConnection;
+import io.goodforgod.testcontainers.extensions.cassandra.ConnectionCassandra;
+import io.goodforgod.testcontainers.extensions.cassandra.Migration;
+import io.goodforgod.testcontainers.extensions.cassandra.TestcontainersCassandra;
+import io.goodforgod.testcontainers.extensions.redis.ConnectionRedis;
+import io.goodforgod.testcontainers.extensions.redis.RedisConnection;
+import io.goodforgod.testcontainers.extensions.redis.TestcontainersRedis;
 import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
@@ -15,36 +18,50 @@ import java.time.Duration;
 import java.util.Map;
 import org.json.JSONObject;
 import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.skyscreamer.jsonassert.JSONAssert;
 import org.skyscreamer.jsonassert.JSONCompareMode;
 
-@TestcontainersPostgreSQL(
+@TestcontainersCassandra(
         network = @Network(shared = true),
         mode = ContainerMode.PER_RUN,
         migration = @Migration(
-                engine = Migration.Engines.FLYWAY,
+                locations = "migrations",
+                engine = Migration.Engines.SCRIPTS,
                 apply = Migration.Mode.PER_METHOD,
                 drop = Migration.Mode.PER_METHOD))
-class PetControllerTests {
+@TestcontainersRedis(
+        network = @Network(shared = true),
+        mode = ContainerMode.PER_RUN)
+class BlackBoxTests {
 
     private static final AppContainer container = AppContainer.build()
             .withNetwork(org.testcontainers.containers.Network.SHARED);
 
-    @ConnectionPostgreSQL
-    private JdbcConnection connection;
+    @ConnectionCassandra
+    private CassandraConnection connection;
 
-    @BeforeAll
-    public static void setup(@ConnectionPostgreSQL JdbcConnection connection) {
-        var params = connection.paramsInNetwork().orElseThrow();
-        container.withEnv(Map.of(
-                "POSTGRES_JDBC_URL", params.jdbcUrl(),
-                "POSTGRES_USER", params.username(),
-                "POSTGRES_PASS", params.password(),
-                "CACHE_EXPIRE_WRITE", "0s"));
+    @BeforeEach
+    public void setup(@ConnectionCassandra CassandraConnection cassandraConnection,
+                      @ConnectionRedis RedisConnection redisConnection) {
+        if (!container.isRunning()) {
+            var paramsCassandra = cassandraConnection.paramsInNetwork().orElseThrow();
+            var paramsRedis = redisConnection.paramsInNetwork().orElseThrow();
+            container.withEnv(Map.of(
+                    "CASSANDRA_CONTACT_POINTS", paramsCassandra.contactPoint(),
+                    "CASSANDRA_USER", paramsCassandra.username(),
+                    "CASSANDRA_PASS", paramsCassandra.password(),
+                    "CASSANDRA_DC", paramsCassandra.datacenter(),
+                    "CASSANDRA_KEYSPACE", paramsCassandra.keyspace(),
+                    "REDIS_URL", paramsRedis.uri().toString(),
+                    "REDIS_USER", paramsRedis.username(),
+                    "REDIS_PASS", paramsRedis.password(),
+                    "CACHE_EXPIRE_WRITE", "0s",
+                    "RETRY_ATTEMPTS", "0"));
 
-        container.start();
+            container.start();
+        }
     }
 
     @AfterAll
@@ -73,7 +90,6 @@ class PetControllerTests {
 
         // then
         connection.assertCountsEquals(1, "pets");
-        connection.assertCountsEquals(1, "categories");
         var responseBody = new JSONObject(response.body());
         assertNotNull(responseBody.query("/id"));
         assertNotEquals(0L, responseBody.query("/id"));
@@ -102,7 +118,6 @@ class PetControllerTests {
         var createResponse = httpClient.send(createRequest, HttpResponse.BodyHandlers.ofString());
         assertEquals(200, createResponse.statusCode(), createResponse.body());
         connection.assertCountsEquals(1, "pets");
-        connection.assertCountsEquals(1, "categories");
         var createResponseBody = new JSONObject(createResponse.body());
 
         // then
@@ -117,6 +132,23 @@ class PetControllerTests {
 
         var getResponseBody = new JSONObject(getResponse.body());
         JSONAssert.assertEquals(createResponseBody.toString(), getResponseBody.toString(), JSONCompareMode.LENIENT);
+    }
+
+    @Test
+    void getPetNotFound() throws Exception {
+        // given
+        var httpClient = HttpClient.newHttpClient();
+
+        // when
+        var getRequest = HttpRequest.newBuilder()
+                .GET()
+                .uri(container.getURI().resolve("/v3/pets/1"))
+                .timeout(Duration.ofSeconds(5))
+                .build();
+
+        // then
+        var getResponse = httpClient.send(getRequest, HttpResponse.BodyHandlers.ofString());
+        assertEquals(404, getResponse.statusCode(), getResponse.body());
     }
 
     @Test
@@ -137,7 +169,6 @@ class PetControllerTests {
         var createResponse = httpClient.send(createRequest, HttpResponse.BodyHandlers.ofString());
         assertEquals(200, createResponse.statusCode(), createResponse.body());
         connection.assertCountsEquals(1, "pets");
-        connection.assertCountsEquals(1, "categories");
         var createResponseBody = new JSONObject(createResponse.body());
 
         // when
@@ -189,7 +220,6 @@ class PetControllerTests {
         var createResponse = httpClient.send(createRequest, HttpResponse.BodyHandlers.ofString());
         assertEquals(200, createResponse.statusCode(), createResponse.body());
         connection.assertCountsEquals(1, "pets");
-        connection.assertCountsEquals(1, "categories");
         var createResponseBody = new JSONObject(createResponse.body());
 
         // when
