+++ kora-java-graalvm-crud-jdbc/src/main/resources/application.conf
@@ -41,15 +41,17 @@ resilient {
     slidingWindowSize = 50
     minimumRequiredCalls = 25
     failureRateThreshold = 50
+    waitDurationInOpenState = "25s"
     permittedCallsInHalfOpenState = 10
-    waitDurationInOpenState = 15s
   }
   timeout.pet {
-    duration = 5000ms
+    duration = "5000ms"
   }
   retry.pet {
-    delay = 100ms
-    attempts = 2
+    delay = "500ms"
+    delayStep = "5s"
+    attempts = 3
+    attempts = ${?RETRY_ATTEMPTS}
   }
 }
