+++ kora-java-graalvm-crud-jdbc/src/test/java/ru/tinkoff/kora/example/graalvm/crud/jdbc/BlackBoxTests.java
@@ -27,7 +27,7 @@ import org.skyscreamer.jsonassert.JSONCompareMode;
                 engine = Migration.Engines.FLYWAY,
                 apply = Migration.Mode.PER_METHOD,
                 drop = Migration.Mode.PER_METHOD))
-class GraalVMPetControllerTests {
+class BlackBoxTests {
 
     private static final AppContainer container = AppContainer.build()
             .withNetwork(org.testcontainers.containers.Network.SHARED);
@@ -43,7 +43,8 @@ class GraalVMPetControllerTests {
                     "POSTGRES_JDBC_URL", params.jdbcUrl(),
                     "POSTGRES_USER", params.username(),
                     "POSTGRES_PASS", params.password(),
-                    "CACHE_EXPIRE_WRITE", "0s"));
+                    "CACHE_EXPIRE_WRITE", "0s",
+                    "RETRY_ATTEMPTS", "0"));
 
             container.start();
         }
@@ -119,6 +120,23 @@ class GraalVMPetControllerTests {
 
         var getResponseBody = new JSONObject(getResponse.body());
         JSONAssert.assertEquals(createResponseBody.toString(), getResponseBody.toString(), JSONCompareMode.LENIENT);
+    }
+
+    @Test
+    void getPetNotFound() throws Exception {
+        // given
+        var httpClient = HttpClient.newHttpClient();
+
+        // when
+        var getRequest = HttpRequest.newBuilder()
+                .GET()
+                .uri(container.getURI().resolve("/v3/pets/1"))
+                .timeout(Duration.ofSeconds(5))
+                .build();
+
+        // then
+        var getResponse = httpClient.send(getRequest, HttpResponse.BodyHandlers.ofString());
+        assertEquals(404, getResponse.statusCode(), getResponse.body());
     }
 
     @Test
