+++ kora-java-graalvm-crud-r2dbc/src/main/java/ru/tinkoff/kora/example/graalvm/crud/r2dbc/controller/HttpExceptionHandler.java
@@ -3,6 +3,8 @@ package ru.tinkoff.kora.example.graalvm.crud.r2dbc.controller;
 import io.micrometer.core.instrument.config.validate.ValidationException;
 import java.util.concurrent.CompletionStage;
 import java.util.concurrent.TimeoutException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import ru.tinkoff.kora.common.Component;
 import ru.tinkoff.kora.common.Context;
 import ru.tinkoff.kora.common.Tag;
@@ -14,6 +16,8 @@ import ru.tinkoff.kora.json.common.JsonWriter;
 @Tag(HttpServerModule.class)
 @Component
 public final class HttpExceptionHandler implements HttpServerInterceptor {
+
+    private static final Logger logger = LoggerFactory.getLogger(HttpExceptionHandler.class);
 
     private final JsonWriter<MessageTO> errorJsonWriter;
 
@@ -35,6 +39,7 @@ public final class HttpExceptionHandler implements HttpServerInterceptor {
             } else if (e instanceof TimeoutException) {
                 return HttpServerResponse.of(408, body);
             } else {
+                logger.error("Request '{} {}' failed", request.method(), request.path(), e);
                 return HttpServerResponse.of(500, body);
             }
         });
