+++ kora-kotlin-crud/src/main/resources/application.conf
@@ -51,6 +51,7 @@ resilient {
     delay = "500ms"
     delayStep = "5s"
     attempts = 3
+    attempts = ${?RETRY_ATTEMPTS}
   }
 }
